+++ src/components/Header.jsx
@@ -1,25 +1,61 @@
 // Copyright (c) 2021 Visiosto oy
 // Licensed under the MIT License
 
-import React, { useContext } from 'react';
+import React from 'react';
 import { useStaticQuery, graphql } from 'gatsby';
+import { getImage, withArtDirection } from 'gatsby-plugin-image';
 import styled from 'styled-components';
 
 import Navigation from './Navigation';
-import ThemeContext from './ThemeContext';
+import SchemedImage from './SchemedImage';
+
+import theme from '../theme';
 
 export default (props) => {
-  const { site } = useStaticQuery(
+  const { site, logoPhoneSLight, logoPhoneSDark, logoTabletLight, logoTabletDark } = useStaticQuery(
     graphql`
       query {
-        site {
+        site: site {
           siteMetadata {
             title
+          }
+        }
+        logoPhoneSLight: file(relativePath: { eq: "header/logo-light.png" }) {
+          childImageSharp {
+            gatsbyImageData(width: 301)
+          }
+        }
+        logoPhoneSDark: file(relativePath: { eq: "header/logo-dark.png" }) {
+          childImageSharp {
+            gatsbyImageData(width: 301)
+          }
+        }
+        logoTabletLight: file(relativePath: { eq: "header/logo-light.png" }) {
+          childImageSharp {
+            gatsbyImageData(width: 301)
+          }
+        }
+        logoTabletDark: file(relativePath: { eq: "header/logo-dark.png" }) {
+          childImageSharp {
+            gatsbyImageData(width: 301)
           }
         }
       }
     `,
   );
+
+  const logosLight = withArtDirection(getImage(logoPhoneSLight), [
+    {
+      media: theme.devices.tablet,
+      image: getImage(logoTabletLight),
+    },
+  ]);
+  const logosDark = withArtDirection(getImage(logoPhoneSDark), [
+    {
+      media: theme.devices.tablet,
+      image: getImage(logoTabletDark),
+    },
+  ]);
 
   const Header = styled.header`
     margin: 2em ${(props) => props.theme.layout.marginPhone};
@@ -34,6 +70,7 @@ export default (props) => {
   `;
 
   const SiteTitle = (props.home ? styled.h1 : styled.p)`
+    display: none;
     margin: 0;
     font-size: 3rem;
     font-family: ${(props) => props.theme.fonts.heading};
@@ -41,9 +78,29 @@ export default (props) => {
     text-align: center;
   `;
 
+  const SiteBranding = styled.div`
+    display: flex;
+    justify-content: center;
+    margin: 2rem auto;
+
+    @media screen and ${(props) => props.theme.devices.tablet} {
+      margin: 3rem auto;
+    }
+  `;
+
+  const Image = styled(SchemedImage)`
+    @media screen and ${(props) => props.theme.devices.tablet} {
+      width: 300px;
+      height: auto;
+    }
+  `;
+
   return (
     <Header>
-      <SiteTitle {...props}>{site.siteMetadata.title}</SiteTitle>
+      <SiteBranding>
+        <SiteTitle {...props}>{site.siteMetadata.title}</SiteTitle>
+        <Image light={logosLight} dark={logosDark} />
+      </SiteBranding>
       <Navigation {...props} />
     </Header>
   );