+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Workflow status badge to the README.
 - Lint for CSS-in-JS.
 - Internationalization of the markdown page links.
+- Company logo to the header of the site.
 
 ### Changed